Extract QScriptCorePlanner from MongoDbPlanner into separate file
